requirements.txt
alabaster>=0.7.10
amqp>=2.1.4
appdirs>=1.4.3
Babel>=2.4.0
billiard>=3.5.0.2
bkcharts>=0.2
bokeh>=0.12.6
boto3>=1.4.4
botocore>=1.5.46
bumpversion>=0.5.3
celery>=4.0.2
certifi>=2017.4.17
chardet>=3.0.4
click>=6.7
cloudpickle>=0.5.2
coreapi>=2.3.3
coreschema>=0.0.4
coverage>=4.4
coveralls>=1.2.0
dask>=1.0.0
distributed>=1.25.0
Django>=2.1.3
django-cleanup>=2.1.0
django-cors-headers>=2.1.0
django-extensions>=1.9.9
djangorestframework>=3.7.7
docopt>=0.6.2
docutils>=0.13.1
HeapDict>=1.0.0
idna>=2.5
imagesize>=0.7.1
itypes>=1.1.0
Jinja2>=2.9.6
jmespath>=0.9.2
kombu>=4.0.2
locket>=0.2.0
MarkupSafe>=1.0
msgpack>=0.5.6
msgpack-python>=0.5.4
nose>=1.3.7
numpy==1.13.0
numpydoc>=0.7.0
packaging>=16.8
pandas>=0.20.1
partd>=0.3.8
psutil>=5.4.3
psycopg2>=2.7.1
py>=1.4.34
pydot>=1.2.3
Pygments>=2.2.0
pyparsing>=2.2.0
pytest>=3.2.3
pytest-cov>=2.5.1
pytest-django>=3.1.2
python-dateutil>=2.6.0
pytz>=2018.3
PyYAML>=3.12
requests>=2.17.3
s3fs>=0.1.0
s3transfer>=0.1.10
scikit-learn>=0.20.0
scipy>=0.19.0
six>=1.11.0
snowballstemmer>=1.2.1
sortedcollections>=0.5.3
sortedcontainers>=1.5.9
Sphinx>=1.6.3
sphinxcontrib-websupport>=1.0.1
tblib>=1.3.2
toolz>=0.9.0
tornado>=4.5.3
typing>=3.6.4
uritemplate>=3.0.0
urllib3>=1.21.1
vine>=1.1.3
zict>=0.1.3
